// src/common/enums/order-status.enum.ts
export enum ConnectionOrderStatus {
  PENDING_ASSIGN = 'PENDING_ASSIGN',
  IN_REVIEW = 'IN_REVIEW',
  WAITING_PURCHASE = 'WAITING_PURCHASE',
  IN_OFFLINE = 'IN_OFFLINE',
  CONFIRMED = 'CONFIRMED',
  CANCELLED = 'CANCELLED',
  FAILED = 'FAILED',
}

export enum EntrustmentOrderStatus {
  PENDING_REVIEW = 'PENDING_REVIEW',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  PROCESSING = 'PROCESSING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

export enum OrderType {
  CONNECTION = 'CONNECTION',
  ENTRUSTMENT = 'ENTRUSTMENT',
}